qvibrateaccessory - use int instead of quint16
It seems that quint16 is not compatible with QVariant, so using integer
we are on the safe side

+++ devices/greenphone/src/plugins/phonevendors/greenphone/vendor_greenphone_p.h
@@ -217,7 +217,7 @@ public:
     ~GreenphoneVibrateAccessory();
 
 public slots:
-    void setVibrateNow( const bool value, quint16 strength = 0xffff, int timeoutMs = 0 );
+    void setVibrateNow( const bool value, int strength = 0xffff, int timeoutMs = 0 );
 };
 
 class GreenphoneServiceNumbers : public QModemServiceNumbers
